Make example in generics.md differentiate type from value
When reading this first time I've found it confusing that `A` is used in places where I'd expect a value. It took me a while to realize that `A` is equivalent to `A{}` which constructs a value of type A.
I think that when explaining a new concept (here generics) it pays of to make everything else as easy to understand as possible.